+++ x-pack/plugin/core/src/main/java/org/elasticsearch/xpack/core/action/GetDataStreamAction.java
@@ -1,64 +1,30 @@
 /*
- * Licensed to Elasticsearch under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
  */
-package org.elasticsearch.action.admin.indices.datastream;
+package org.elasticsearch.xpack.core.action;
 
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-import org.elasticsearch.action.ActionListener;
 import org.elasticsearch.action.ActionRequestValidationException;
 import org.elasticsearch.action.ActionResponse;
 import org.elasticsearch.action.ActionType;
 import org.elasticsearch.action.IndicesRequest;
-import org.elasticsearch.action.support.ActionFilters;
 import org.elasticsearch.action.support.IndicesOptions;
 import org.elasticsearch.action.support.master.MasterNodeReadRequest;
-import org.elasticsearch.action.support.master.TransportMasterNodeReadAction;
 import org.elasticsearch.cluster.AbstractDiffable;
-import org.elasticsearch.cluster.ClusterState;
-import org.elasticsearch.cluster.block.ClusterBlockException;
-import org.elasticsearch.cluster.block.ClusterBlockLevel;
 import org.elasticsearch.cluster.health.ClusterHealthStatus;
-import org.elasticsearch.cluster.health.ClusterStateHealth;
 import org.elasticsearch.cluster.metadata.DataStream;
-import org.elasticsearch.cluster.metadata.IndexNameExpressionResolver;
-import org.elasticsearch.cluster.metadata.MetadataIndexTemplateService;
-import org.elasticsearch.cluster.service.ClusterService;
 import org.elasticsearch.common.Nullable;
 import org.elasticsearch.common.ParseField;
-import org.elasticsearch.common.inject.Inject;
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
-import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.common.xcontent.ToXContentObject;
 import org.elasticsearch.common.xcontent.XContentBuilder;
-import org.elasticsearch.index.Index;
-import org.elasticsearch.threadpool.ThreadPool;
-import org.elasticsearch.transport.TransportService;
 
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Comparator;
 import java.util.List;
-import java.util.Map;
 import java.util.Objects;
-import java.util.stream.Collectors;
 
 public class GetDataStreamAction extends ActionType<GetDataStreamAction.Response> {
 
@@ -75,6 +41,10 @@ public class GetDataStreamAction extends ActionType<GetDataStreamAction.Response
 
         public Request(String[] names) {
             this.names = names;
+        }
+
+        public String[] getNames() {
+            return names;
         }
 
         @Override
@@ -259,64 +229,6 @@ public class GetDataStreamAction extends ActionType<GetDataStreamAction.Response
         @Override
         public int hashCode() {
             return Objects.hash(dataStreams);
-        }
-    }
-
-    public static class TransportAction extends TransportMasterNodeReadAction<Request, Response> {
-
-        private static final Logger logger = LogManager.getLogger(TransportAction.class);
-
-        @Inject
-        public TransportAction(TransportService transportService, ClusterService clusterService, ThreadPool threadPool,
-                               ActionFilters actionFilters, IndexNameExpressionResolver indexNameExpressionResolver) {
-            super(NAME, transportService, clusterService, threadPool, actionFilters, Request::new, indexNameExpressionResolver);
-        }
-
-        @Override
-        protected String executor() {
-            return ThreadPool.Names.SAME;
-        }
-
-        @Override
-        protected Response read(StreamInput in) throws IOException {
-            return new Response(in);
-        }
-
-        @Override
-        protected void masterOperation(Request request, ClusterState state,
-                                       ActionListener<Response> listener) throws Exception {
-            List<DataStream> dataStreams = getDataStreams(state, indexNameExpressionResolver, request);
-            List<Response.DataStreamInfo> dataStreamInfos = new ArrayList<>(dataStreams.size());
-            for (DataStream dataStream : dataStreams) {
-                String indexTemplate = MetadataIndexTemplateService.findV2Template(state.metadata(), dataStream.getName(), false);
-                String ilmPolicyName = null;
-                if (indexTemplate != null) {
-                    Settings settings = MetadataIndexTemplateService.resolveSettings(state.metadata(), indexTemplate);
-                    ilmPolicyName = settings.get("index.lifecycle.name");
-                } else {
-                    logger.warn("couldn't find any matching template for data stream [{}]. has it been restored (and possibly renamed)" +
-                        "from a snapshot?", dataStream.getName());
-                }
-                ClusterStateHealth streamHealth = new ClusterStateHealth(state,
-                    dataStream.getIndices().stream().map(Index::getName).toArray(String[]::new));
-                dataStreamInfos.add(new Response.DataStreamInfo(dataStream, streamHealth.getStatus(), indexTemplate, ilmPolicyName));
-            }
-            listener.onResponse(new Response(dataStreamInfos));
-        }
-
-        static List<DataStream> getDataStreams(ClusterState clusterState, IndexNameExpressionResolver iner, Request request) {
-            List<String> results = iner.dataStreamNames(clusterState, request.indicesOptions(), request.names);
-            Map<String, DataStream> dataStreams = clusterState.metadata().dataStreams();
-
-            return results.stream()
-                .map(dataStreams::get)
-                .sorted(Comparator.comparing(DataStream::getName))
-                .collect(Collectors.toList());
-        }
-
-        @Override
-        protected ClusterBlockException checkBlock(Request request, ClusterState state) {
-            return state.blocks().globalBlockedException(ClusterBlockLevel.METADATA_WRITE);
         }
     }
