+++ x-pack/plugin/data-streams/src/test/java/org/elasticsearch/xpack/core/action/DataStreamsStatsResponseTests.java
@@ -1,23 +1,9 @@
 /*
- * Licensed to Elasticsearch under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
  */
-
-package org.elasticsearch.action.admin.indices.datastream;
+package org.elasticsearch.xpack.core.action;
 
 import org.elasticsearch.ElasticsearchException;
 import org.elasticsearch.action.support.DefaultShardOperationFailedException;
@@ -53,20 +39,33 @@ public class DataStreamsStatsResponseTests extends AbstractWireSerializingTestCa
             long storeSize = randomLongBetween(250, 1000000000);
             totalStoreSize += storeSize;
             long maximumTimestamp = randomRecentTimestamp();
-            dataStreamStats.add(new DataStreamsStatsAction.DataStreamStats(dataStreamName, backingIndices,
-                new ByteSizeValue(storeSize), maximumTimestamp));
+            dataStreamStats.add(
+                new DataStreamsStatsAction.DataStreamStats(dataStreamName, backingIndices, new ByteSizeValue(storeSize), maximumTimestamp)
+            );
         }
         int totalShards = randomIntBetween(backingIndicesTotal, backingIndicesTotal * 3);
         int successfulShards = randomInt(totalShards);
         int failedShards = totalShards - successfulShards;
         List<DefaultShardOperationFailedException> exceptions = new ArrayList<>();
         for (int i = 0; i < failedShards; i++) {
-            exceptions.add(new DefaultShardOperationFailedException(randomAlphaOfLength(8).toLowerCase(Locale.getDefault()),
-                randomInt(totalShards), new ElasticsearchException("boom")));
+            exceptions.add(
+                new DefaultShardOperationFailedException(
+                    randomAlphaOfLength(8).toLowerCase(Locale.getDefault()),
+                    randomInt(totalShards),
+                    new ElasticsearchException("boom")
+                )
+            );
         }
-        return new DataStreamsStatsAction.Response(totalShards, successfulShards, failedShards, exceptions,
-            dataStreamCount, backingIndicesTotal, new ByteSizeValue(totalStoreSize),
-            dataStreamStats.toArray(new DataStreamsStatsAction.DataStreamStats[0]));
+        return new DataStreamsStatsAction.Response(
+            totalShards,
+            successfulShards,
+            failedShards,
+            exceptions,
+            dataStreamCount,
+            backingIndicesTotal,
+            new ByteSizeValue(totalStoreSize),
+            dataStreamStats.toArray(new DataStreamsStatsAction.DataStreamStats[0])
+        );
     }
 
     private static long randomRecentTimestamp() {
