+++ x-pack/plugin/ml/qa/native-multi-node-tests/build.gradle
@@ -8,6 +8,7 @@ dependencies {
   testImplementation project(path: xpackModule('ml'))
   testImplementation project(path: xpackModule('ml'), configuration: 'testArtifacts')
   testImplementation project(path: ':modules:ingest-common')
+  testImplementation project(path: xpackModule('data-streams'))
 }
 
 // location for keys and certificates
